1. Fix bug on header ordering. Now, before request a header item, the mask field of HDITEM is filled with proper flags and text of the header index is filled again before update the header item (according to microsoft docs).  2. Duration compare function now order items in correct order. 3. Add a missing break on  WA_UI_Listview_ShowItemContextMenu 4. Now main window don't show the existing window of an existing instance when a file is opened from the windows explorer

+++ WinAudio/SearchDlg.c
@@ -128,7 +128,7 @@ INT_PTR CALLBACK SearchDlgProc(HWND hDlg, UINT message, WPARAM wParam, LPARAM lP
 
                 if (nSelectedItem != LB_ERR)
                 {
-                    nItemIndex = ListBox_GetItemData(GetDlgItem(hDlg, IDC_SEARCH_DLG_LIST), nSelectedItem);
+                    nItemIndex = (int32_t) ListBox_GetItemData(GetDlgItem(hDlg, IDC_SEARCH_DLG_LIST), nSelectedItem);
 
                     if (nItemIndex != LB_ERR)
                     {
